pipeline/requirements.txt
pytest>=8.0
requests>=2.31
aiohttp>=3.9
SPARQLWrapper>=2.0
zeep>=4.2
python-dotenv>=1.0
thefuzz>=0.22
# rdkit-pypi>=2024.3.1  # Optional: InChI matching (install manually if needed)
